Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/test-expected.txt

Horizontal grid and verticalLR item with relative width
É É ÉÉ

FAIL .inline-grid 1 assert_equals: 
<div class="inline-grid justifyItemsBaseline" data-expected-width="180" data-expected-height="300">
  <div class="firstRowFirstColumn fixedWidth" data-offset-x="30" data-offset-y="0" data-expected-width="125" data-expected-height="100"></div>
  <div class="secondRowFirstColumn bigFont paddingLeft verticalLR" data-offset-x="0" data-offset-y="100" data-expected-width="120" data-expected-height="200">É É ÉÉ</div>
  <div class="autoRowSpanning2AutoColumn width25"></div>
</div>
width expected 180 but got 150
